make dse headers into dse.h
