--- platform/dse/include/sched.h
@@ -1,54 +1,0 @@
-/****************************************************************************
- * Copyright (c) 2012, the Konoha project authors. All rights reserved.
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- *  * Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *  * Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
- * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
- * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
- * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- ***************************************************************************/
-
-/* ************************************************************************ */
-
-#ifndef SCHED_H_
-#define SCHED_H_
-
-#ifndef K_USE_PTHREAD
-#include <pthread.h>
-#endif /* K_USE_PTHREAD */
-
-#include "message.h"
-
-#define MSG_QUEUE_SIZE 16
-#define next(index) (((index) + 1) % MSG_QUEUE_SIZE)
-
-struct Scheduler {
-	Message *msgQueue[MSG_QUEUE_SIZE];
-	int front;
-	int last;
-	pthread_mutex_t lock;
-	pthread_cond_t cond;
-};
-
-typedef struct Scheduler Scheduler;
-
-Scheduler *Scheduler_new(void);
-void Scheduler_delete(Scheduler *sched);
-bool dse_enqueue(Scheduler *sched, Message *msg);
-Message *dse_dequeue(Scheduler *sched);
-
-#endif /* SCHED_H_ */
