+++ platform/dse/src/util.c
@@ -24,7 +24,7 @@
 
 /* ************************************************************************ */
 
-#include "util.h"
+#include "dse.h"
 
 size_t totalMalloc = 0;
